Fix display record info in RecordsList (date format)

+++ app/src/main/java/el/ka/dictophone/adapters/RecordsAdapter.kt
@@ -28,8 +28,9 @@ class RecordsAdapter(
             itemView.setOnClickListener {
                 onItemClickListener(record)
             }
+
             tvName.text = record.name
-            tvDate.text = sdf.format(Date(record.createAt.toLong()))
+            tvDate.text = sdf.format(Date(record.createAt))
             tvDuration.text = record.duration.toString()
         }
     }
