+++ app/src/main/java/el/ka/dictophone/models/Record.kt
@@ -4,6 +4,6 @@ data class Record(
     var id: Int,
     var name: String,
     var filePath: String,
-    var createAt: Int,
+    var createAt: Long,
     var duration: Int
 )
